Removed the period param, as it is not needed in this sdk
The period parameter is not needed, as metrics do not need to be aggregated over a period. Thus I removed it from all client facing code and for usage of the cloudwatch api I added a constant period of 86400 seconds (1 day) to keep cost low.

Moreover I also added first pre-commit hooks to the repo that will be executed locally on commit.